[Fontations] Fix Rust edition, do not require extern crate
With the explanations in [1], fix resolution of Rust crate
dependencies by providing Meson with information that 2021 is the
Rust edition to use.

This fixes two issues: We no longer need the extern crate... lines
in fc-fontations/mod.rs, and this resolves dependency problems
when using meson 1.8.0 and bytemuck-derive depending on syn.

[1] https://github.com/mesonbuild/meson/issues/13753#issuecomment-2394010571