Do not fallback decoding with UTF-16BE if no iconv support
This fallback was added to decode a string in a font that doesn't have
correct encoding ID in the name table for some reasons though,
it doesn't work effectively without iconv support.  Even if it has
correct sequences for a certain encoding, incorrect string is decoded
because of missing correct decoder.  So let's drop this fallback if
iconv support is disabled.

Fixes https://gitlab.freedesktop.org/fontconfig/fontconfig/-/issues/479

Changelog: changed