Add fontconfig version in FcCache
Even if cache files needs to be updated, do not re-generate them
if it was generated by newer version of fontconfig.

Fontconfig cache files are basically compatible with older versions
of fontconfig unless any drastic changes happens there and bump a
cache version. older versions of fontconfig simply ignore unknown
objects in caches.
However, a troublesome situation is coming by sharing caches between
host and containers. A situation of updating caches is likely to
happens on even containers when mtime of font directories are updated.
And it may drops such unknown objects. So this aims to avoid such
situation and keep not updating and warn to encourage users to update
caches at host.

Also, added FONTCONFIG_NO_CHECK_CACHE_VERSION environment variable to
revert this behavior for contigency plan. This will be removed in
the future.

Changelog: changed