[Fontations] Fix use-after-free in handling exclusive lang
The exclusive_lang() function returns an Option<CString> which needs
to be retained until the call to FcLangSetFromCharSet is completed.
The CString owns the string memory, which we pass to FcLangSetFromCharSet.

Previously, the code dropped the Option<CString> during the conversion
to a char pointer, thus returning a pointer to dropped memory.

Identified by ASAN in Chromium.