Fix memory leak after path canonicalization in e42188283f0ee
FcValueSave (called from FcPatternAddString) calls FcStrDup for incoming
string objects, which means incoming strings are duplicated and owned when
saving. This means, callers need to free passed strings.

For canon_file this means, it needs to be freed not only in the erro
case but also in the regular case, similar to foundry_.

Introduced in e42188283f0ee1fb23089f16dbf95e0f3bcbacaf.