[Fontations] Don't leak PatternElement values
FcPatternObjectAdd(Range|CharSet|LangSet) do clone the values,
so we need to free them on the Rust side and not leak them.

However, the FcFontSetAdd() call does not clone the pattern
that is passed in, so there, we do still need to use .into_raw().

This leak was identified in a downstream Chromium ASAN
test build when using Fontations indexing. Fix manually
verified against this test.