[Fontations] Match FreeType for two types of broken fonts
When rolling out Fontations indexing in Chrome, we encountered
issues in name matching when Chromium came across broken fonts in
indexing.

For fonts with a missing family name, FreeType generates a family name
from the file name, implement that in Fontations.

For fonts with a broken major version in the CFF table, FreeType rejects
the file and does not create a Pattern entry, match that in
Fontations.

Add two test fonts that cover these situations, derived from Noto Sans
CJK and Robot respectively, manually patched to not have a family name
and to have an incorrect CFF major version of 4.

Add FreeType/Fontations comparison testing to allow for a mode where
return codes must match so that a failed fc-query run can be compared,
too.

Match normalization of file paths to FreeType in Fontations indexing.