detect mkostemp with stdlib.h
On Apple platforms it may be detected as available by
just linking with it.
But when calling it it may be turned into a weak
reference that will fail to run on older platforms.

It's possible to get around this using __builtin_available() [^1]
but due to the way this code is used a clean fallback is not
possible.

It's available since macOS 10.12, iOS 10.0, tvOS 10.0 and watchOS 3.0.

[^1]: https://clang.llvm.org/docs/LanguageExtensions.html#objective-c-available