Add fc-genconf the configuration generator tool
A primitive tool to genarete a configuration in XML format
from the font metadata in a font file.
Supporting few optional parameters to determine a generic
family name and languages to make it conditional for.
In most cases, it will needs to be set properly at the moment
otherwise the generated configuration won't be functional.

Just for a first step.

Changelog: added