Minimal preparations for FreeType-less build
Preparations for a FreeType-less build. Default enable ENABLE_FREETYPE
compile time define. Introduce conditional compilation in required core
library source files. This is currently not intended as a configurable
option in the meson or autoconf builds, but only intended for unblocking
compilation of the FontConfig library without the FreeType dependency.

This is the minimal required modification for downstream projects to
build the FontConfig library without FreeType.

Partially addresses #502.