Fix unexpected priority change when looking up by specific family name
"prepend"ing an item without certain lookup for the targeted object
is always adding at the top of elements. This makes it highest priority
and is not what this config expects. Also, there is the order issue
when multiple values in genericfamily is available and it is hard to
address in current configuration scheme.
So looking up first genericfamily only to workaround and append it
to complement family.